Address review comments and move snippets to a single file
