+++ modules/gapi/include/opencv2/gapi/s11n/base.hpp
@@ -40,7 +40,7 @@ struct NotImplemented {
  * which can be utilized when serializing a custom type.
  *
  * Example of usage:
- * @snippet modules/gapi/samples/custom_type_serialization.cpp S11N usage
+ * @snippet modules/gapi/samples/api_ref_snippets.cpp S11N usage
  *
  */
 template<typename T>
